Update .gitignore
1) re folder changes in build process
2) to ignore .res files

Old build process created binary files in Build directory while new
build process uses _build. So Build was removed from .gitignore and
_build was added.

There is a possibility that a .res file may be created in the Src folder
that would be listed by Git as a new file. This was not wanted.

FIXUP .gitignore